Added script to detect a sudoku playing field
The commited script is a first attempt on a script
that detects a sudoku playing field on a given picture.
This script uses Hough Line Transformation to detect the
lines of the playing field.
As for now only the detection of the relevant lines is implemented,
the extraction of the playing field is still to do.